CV-Platform/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
asyncpg
psycopg2-binary
passlib[bcrypt]
python-jose